Добавлена проверка на активность кеша, т.к. модуль Cache не проверяет активность кеша Kohana, поэтому нужно вручную делать проверку.

+++ cms/modules/kodicms/classes/kodicms/controller/api/cache.php
@@ -14,10 +14,13 @@ class KodiCMS_Controller_API_Cache extends Controller_System_Api {
 	
 	public function get_clear()
 	{
-		Cache::instance()->delete_all();
-		Kohana::cache('Kohana::find_file()', NULL, -1);
-		Kohana::cache('Route::cache()', NULL, -1);
-		Kohana::cache('profiler_application_stats', NULL, -1);
+		if(Kohana::$caching === TRUE)
+		{
+			Cache::instance()->delete_all();
+			Kohana::cache('Kohana::find_file()', NULL, -1);
+			Kohana::cache('Route::cache()', NULL, -1);
+			Kohana::cache('profiler_application_stats', NULL, -1);
+		}
 		
 		Kohana::$log->add(Log::INFO, ':user clear cache')->write();
