+++ cms/modules/plugins/classes/plugin/decorator.php
@@ -345,13 +345,16 @@ class Plugin_Decorator extends Plugin {
 	 */
 	protected function _clear_cache()
 	{
-		$cache = Cache::instance();
-		
-		$cache->delete('Database::cache('.self::CACHE_KEY . '::list)');
-		$cache->delete('Database::cache('.self::CACHE_KEY . '::plugin::' . $this->id() . ')');
-		
-		Kohana::cache('Route::cache()', NULL, -1);
-		Kohana::cache('Kohana::find_file()', NULL, -1);
+		if(Kohana::$caching === TRUE)
+		{
+			$cache = Cache::instance();
+
+			$cache->delete('Database::cache('.self::CACHE_KEY . '::list)');
+			$cache->delete('Database::cache('.self::CACHE_KEY . '::plugin::' . $this->id() . ')');
+
+			Kohana::cache('Route::cache()', NULL, -1);
+			Kohana::cache('Kohana::find_file()', NULL, -1);
+		}
 
 		return $this;
 	}
